[powerpc] Fix transactional FP/VMX/VSX unavailable handlers
Message-id: <20140204154720.21282.7232.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 75967
O-Subject: [PATCH RHEL7.0 BZ1052083 16/9] powerpc: Fix transactional FP/VMX/VSX unavailable handlers
Bugzilla: 1052083
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Seth Jennings <sjenning@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1052083

Description:
------------
Currently, if a process starts a transaction and then takes an
exception because the FPU, VMX or VSX unit is unavailable to it,
we end up corrupting any FP/VMX/VSX state that was valid before
the interrupt.  For example, if the process starts a transaction
with the FPU available to it but VMX unavailable, and then does
a VMX instruction inside the transaction, the FP state gets
corrupted.

Loading up the desired state generally involves doing a reclaim
and a recheckpoint.  To avoid corrupting already-valid state, we have
to be careful not to reload that state from the thread_struct
between the reclaim and the recheckpoint (since the thread_struct
values are stale by now), and we have to reload that state from
the transact_fp/vr arrays after the recheckpoint to get back the
current transactional values saved there by the reclaim.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
7.0

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=3ac8ff1c475bda7174fce63230c0932454287cd5

----------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------

Signed-off-by: Jarod Wilson <jarod@redhat.com>