[redhat] kernel.spec: use 3072-bit RSA module signatures
Message-id: <20140204160918.GB2891@redhat.com>
Patchwork-id: 75968
O-Subject: [RHEL7 PATCH] redhat/kernel.spec: use 3072-bit RSA module signatures
Bugzilla: 1051579
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>

Bugzilla 1051579

RSA signatures of any bit size but 2048 and 3072 are not permitted in FIPS
mode, so we have to change from 4096 down to 3072 bits to be fully
FIPS-compliant. The signatures are used for integrity checks for all
modules upon load, signature verification failure is a panic in FIPS mode.

Signed-off-by: Jarod Wilson <jarod@redhat.com>

---

Signed-off-by: Jarod Wilson <jarod@redhat.com>