[powerpc] Fix 32-bit frames for signals delivered when transactional
Message-id: <20140204162700.21676.57443.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 75970
O-Subject: [PATCH RHEL7.0 BZ1059703] powerpc: Fix 32-bit frames for signals delivered when transactional
Bugzilla: 1059703
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1059703

Description:
------------
Commit d31626f70b61 ("powerpc: Don't corrupt transactional state when
using FP/VMX in kernel") introduced a bug where the uc_link and uc_regs
fields of the ucontext_t that is created to hold the transactional
values of the registers in a 32-bit signal frame didn't get set
correctly.  The reason is that we now clear the MSR_TS bits in the MSR
in save_tm_user_regs(), before the code that sets uc_link and uc_regs.
To fix this, we move the setting of uc_link and uc_regs into the same
if statement that selects whether to call save_tm_user_regs() or
save_user_regs().

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
RHEL 7.0

Brew:
-----
Jarod: This patch is built on top of the patchset here
http://post-office.corp.redhat.com/archives/rhkernel-list/2014-January/msg02933.html

http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6997978

Upstream:
---------
http://git.kernel.org/cgit/linux/kernel/git/benh/powerpc.git/commit/?h=next&id=d765ff23e3181413fb1bed090c8d702165448a84

Test Status:
------------
Madhavan Srinivasan helped me test this.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------

Signed-off-by: Jarod Wilson <jarod@redhat.com>