[scsi] hpsa: mark hpsa as a tech preview
Message-id: <52F37FB2.7060107@redhat.com>
Patchwork-id: 76029
O-Subject: [RHEL7 e-stor PATCH] hpsa: mark hpsa as a tech preview
Bugzilla: 1061210
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Maurizio Lombardi <mlombard@redhat.com>

The hpsa allows with the hpsa_allow_any option that the driver can be used
together with the old cciss devices (cciss has been removed from RHEL7)
and also with new cards not yet added to the pci-id table. With this patch
we cover both cases with a message about a not supported controller.
"TECH PREVIEW: hpsa: this controller may not be fully supported."
"Please review provided documentation for limitations."

upstream - rh specific

Signed-off-by: Jarod Wilson <jarod@redhat.com>