[fs] gfs2: Wait for async DIO in glock state changes
Message-id: <51412ba9082d4369b8759c136e21f73e591c9005.1391782907.git.rpeterso@redhat.com>
Patchwork-id: 76044
O-Subject: [RHEL7.0 PATCH 2/4] GFS2: Wait for async DIO in glock state changes
Bugzilla: 1052927
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

From: Steven Whitehouse <swhiteho@redhat.com>

We need to wait for any outstanding DIO to complete in a couple
of situations. Firstly, in case we are changing out of deferred
mode (in inode_go_sync) where GLF_DIRTY will not be set. That
call could be prefixed with a test for gl_state == LM_ST_DEFERRED
but it doesn't seem worth it bearing in mind that the test for
outstanding DIO is very quick anyway, in the usual case that there
is none.

The second case is in inode_go_lock which will catch the cases
where we have a cached EX lock, but where we grant deferred locks
against it so that there is no glock state transistion. We only
need to wait if the state is not deferred, since DIO is valid
anyway in that state.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>

rhbz#1052927
Upstream-id: 582d2f7aedfde9e1a01170deb003df510aa778d3

Signed-off-by: Jarod Wilson <jarod@redhat.com>