[fs] gfs2: journal data writepages update
Message-id: <299703031.1761895.1392128566655.JavaMail.zimbra@redhat.com>
Patchwork-id: 76119
O-Subject: [RHEL7.0 PATCH] GFS2: journal data writepages update
Bugzilla: 1010452
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch is for bz#1010452 - GFS2: Data corruption when using O_SYNC w/ jdata

The upstream version of the patch may be found here:
http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs?id=774016b2d455017935b3e318b6cc4e055e9dd47f

It was tested by Cluster QE and found to fix the problem.

I built this (and other patches) into a kernel with this build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=7013418

Patch text:

GFS2: journal data writepages update

GFS2 has carried what is more or less a copy of the
write_cache_pages() for some time. It seems that this
copy has slipped behind the core code over time. This
patch brings it back uptodate, and in addition adds the
tracepoint which would otherwise be missing.

We could go further, and eliminate some or all of the
code duplication here. The issue is that if we do that,
then the function we need to split out from the existing
write_cache_pages(), which will look a lot like
gfs2_jdata_write_pagevec(), would land up putting quite a
lot of extra variables on the stack. I know that has been
a problem in the past in the writeback code path, which
is why I've hesitated to do it here.

rhbz#1010452

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>