[fs] nfsd4: fix delegation-unlink/rename race
Message-id: <1392305757-5658-18-git-send-email-bfields@redhat.com>
Patchwork-id: 76802
O-Subject: [Patch RHEL7 17/25] nfsd4: fix delegation-unlink/rename race
Bugzilla: 1031097
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

If a file is unlinked or renamed between the time when we do the local
open and the time when we get the delegation, then we will return to the
client indicating that it holds a delegation even though the file no
longer exists under the name it was open under.

But a client performing an open-by-name, when it is returned a
delegation, must be able to assume that the file is still linked at the
name it was opened under.

So, hold the parent i_mutex for longer to prevent concurrent renames or
unlinks.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream 4335723e8e9fdc6e4bb2555696bc7f1abe75f200
Bugzilla 1031097

Signed-off-by: Jarod Wilson <jarod@redhat.com>