[infiniband] mlx4: Make sure GID index 0 is always occupied
Message-id: <7e601d44ec1da07d7b6f878a36eeda510d867a60.1392321385.git.dledford@redhat.com>
Patchwork-id: 76842
O-Subject: [Patch RHEL7 2/6] IB/mlx4: Make sure GID index 0 is always occupied
Bugzilla: 1061729
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Moni Shoua <monis@mellanox.co.il>

Make sure that for Ethernet ports, the port GID table index 0 is always
occupied with a default GID of the relevant IPv6 link-local adderss.

This provides better user-experience for legacy applications that don't use
the rdma-cm and where working on index 0 prior to the IP addressing change.

Also, as GIDs are generated from IP addresses of the network devices that
are associated with the port, its basically possible that the GID table
will be empty if no IP address was assigned. This doesn't comply to the
IB spec section 4.1.1 "GID usage and properties".

Signed-off-by: Moni Shoua <monis@mellanox.co.il>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>