[fs] cifs: ensure that uncached writes handle unmapped areas correctly
Message-id: <1392382402-2209-2-git-send-email-sprabhu@redhat.com>
Patchwork-id: 76861
O-Subject: [RHEL7 PATCH 1/2] [bz1062588]cifs: ensure that uncached writes handle unmapped areas correctly
Bugzilla: 1062588
CVE: CVE-2014-0069
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Backported from patch in bz 1062578.
Reproducer tested against a kernel containing this patch.

Original summary follows
--
It's possible for userland to pass down an iovec via writev() that has a
bogus user pointer in it. If that happens and we're doing an uncached
write, then we can end up getting less bytes than we expect from the
call to iov_iter_copy_from_user.

cifs_iovec_write isn't set up to handle that situation however. It'll
blindly keep chugging through the page array and not filling those pages
with anything useful. Worse yet, we'll later end up with a negative
number in wdata->tailsz, which will confuse the sending routines and
cause an oops at the very least.

Fix this by having the copy phase of cifs_iovec_write stop copying data
in this situation and send the last write as a short one. At the same
time, we want to avoid sending a zero-length write to the server, so
break out of the loop and set rc to -EFAULT if that happens. This also
allows us to handle the case where no address in the iovec is valid.

Reviewed-by: Pavel Shilovsky <piastry@etersoft.ru>
Reported-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
--
The patch has been recently sent upstream and hence doesn't have an
upstream commit yet.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>