[fs] cifs: sanity check length of data to send before sending
Message-id: <1392382402-2209-3-git-send-email-sprabhu@redhat.com>
Patchwork-id: 76860
O-Subject: [RHEL7 PATCH 2/2] [bz1062588] cifs: sanity check length of data to send before sending
Bugzilla: 1062588
CVE: CVE-2014-0069
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Compiled with and tested using the reproducer in bz 106578.

Original description follows
----
We had a bug discovered recently where an upper layer function
(cifs_iovec_write) could pass down a smb_rqst with an invalid amount of
data in it. The length of the SMB frame would be correct, but the rqst
struct would cause smb_send_rqst to send nearly 4GB of data.

This should never be the case. Add some sanity checking to the beginning
of smb_send_rqst that ensures that the amount of data we're going to
send agrees with the length in the RFC1002 header. If it doesn't, WARN()
and return -EIO to the upper layers.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
---

This has just been sent upstream and hence no upstream commit exists
yet.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>