[kernel] lockdep: increase MAX_LOCKDEP_ENTRIES
Message-id: <20140217195022.GA27860@redacted.bos.redhat.com>
Patchwork-id: 76969
O-Subject: [RHEL7 PATCH] lockdep: increase MAX_LOCKDEP_ENTRIES
Bugzilla: 989385
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=989385

Doubling MAX_LOCKDEP_ENTRIES results in a static memory usage
increase by lockdep from ~900K to ~1.8M, but it will only effect
debug kernels which have substancially reduced performance
and increased memory usage anyway (numbers for x86_64.) In any case,
lockdep is getting disabled due to running out of space when
testing with debug kernels, so we need to increase it if we want
coverage.

As with RHEL6, in some circumstances, the upstream value of
MAX_LOCKDEP_ENTRIES is too low to work in all circumstances when running
KernelTier1 tests. So let's double it as we did in RHEL6, commit
818937af5c6 (RHEL6: lockdep: double MAX_LOCKDEP_ENTRIES).

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>

---

I'll investigate making this somewhat more robust. Possibly as a cmdline
option to the kernel upstream. But for now, this is needed to get some
kernel-debug tests to pass.

Signed-off-by: Jarod Wilson <jarod@redhat.com>