[x86] irq: Fix kbuild warning in smp_irq_move_cleanup_interrupt()
Message-id: <1393587497-23206-1-git-send-email-prarit@redhat.com>
Patchwork-id: 77691
O-Subject: [RHEL7.0 PATCH BZ 1065735] x86/irq: Fix kbuild warning in smp_irq_move_cleanup_interrupt()
Bugzilla: 1065735
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Team,

Current upstream has a bug in which the irq release thread
smp_irq_move_cleanup_interrupt() is not running on all cpus.  The
end result is so similar to what is happening in RHEL7 that I linked the
two behaviours and thought they were the same.  In RHEL7 the issue is that
this calculation below is busted because of the use of cast to unsigned
int.  It turns out the upstream issue and the RHEL7 are not the same and
this patch I submitted previously fixes the issue in RHEL7.  I'm currently
helping some of the Intel folks debug the upstream issue (we're pretty
sure we have identified what is broken and why it is broken).

This took much longer to debug because of that ...

P.

----8<----

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1065735
Build OS: Red Hat Enterprise Linux Server release 7.0 Beta (Maipo)
System name: intel-canoepass-11.lab.bos.redhat.com with -j32
Built on: kernel-3.10.0-95.el7
Arch built: ppc64 s390x x86_64

This fixes an issue where IRQs are not released when they are freed.
This took more time than it should have to narrow down because upstream
currently has a similar bug but it is not exactly the same.

commit c7a730fa4624092e2d1c0cb7b750816e87c32364
Author: Prarit Bhargava <prarit@redhat.com>
Date:   Mon Jan 13 08:40:20 2014 -0500

    x86/irq: Fix kbuild warning in smp_irq_move_cleanup_interrupt()

    Fengguang Wu's 0day kernel build service reported the following build warning:

      arch/x86/kernel/apic/io_apic.c:2211
      smp_irq_move_cleanup_interrupt() warn: always true condition '(irq <= -1) => (0-u32max <= (-1))'

    because irq is defined as an unsigned int instead of an int.
    Fix this trivial error by redefining irq as a signed int.  The
    remaining consumers of the int are okay.

    Signed-off-by: Prarit Bhargava <prarit@redhat.com>
    Cc: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
    Cc: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>
    Cc: Joerg Roedel <joro@8bytes.org>
    Cc: Fengguang Wu <fengguang.wu@intel.com>
    Link: http://lkml.kernel.org/r/1389620420-7110-1-git-send-email-prarit@redhat.com
    Signed-off-by: Ingo Molnar <mingo@kernel.org>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1065735

Cc: Igor Mammedov <imammedo@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: lszubowi@redhat.com
Cc: tcamuso@redhat.com
Signed-off-by: Jarod Wilson <jarod@redhat.com>