[fs] Revert: xfs: reinstate the ilock in xfs_readdir
Message-id: <1393875041-61388-2-git-send-email-bfoster@redhat.com>
Patchwork-id: 77762
O-Subject: [PATCH 1/2] Revert "[fs] xfs: reinstate the ilock in xfs_readdir"
Bugzilla: 1068855

This reverts the following rhel7 commit:

  2017f69a [fs] xfs: reinstate the ilock in xfs_readdir

This commit was introduced upstream to address coherency issues
for CXFS and pulled back into RHEL as part of a major subsystem
update. Soak time has revealed that this patch introduces various
readdir lockdep problems and may need to be reworked upstream.
As CXFS is not a use case for RHEL, revert this commit until a more
robust solution is worked out upstream.

Signed-off-by: Brian Foster <bfoster@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>