[md] dm-snapshot: fix data corruption
Message-id: <alpine.LRH.2.02.1403051157370.27931@file01.intranet.prod.int.rdu2.redhat.com>
Patchwork-id: 77927
O-Subject: [PATCH RHEL7] bz1070291 dm snapshot: fix data corruption
Bugzilla: 1070291
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

Hi

This patch fixes the bug 1070291.

Tests in the bugzilla were run to verify that the patch fixes the data
corruption.

Upstream status - not yet upstream, the patch is expected to be sent
upstream with the next batch of patches for 3.14-rc.

The patch 55494bf2947dccdf2d98b62374fea7365dfead84 broke snapshots. When
loading the list of exceptions, the old code read exception areas one by
one into ps->area and inserted the exceptions into the hash table. The
patch 55494bf2 changed it so that dm-bufio with prefetch is used to load
exceptions. It loads the exceptions correctly, but leaves ps->area
uninitialized. When a new exception is allocated, it is stored in this
uninitialized ps->area location and full ps->area is written to the disk.
This causes metadata corruption.

This patch fixes the bug by copying the last area that was read via
dm-bufio into ps->area.

Upstream linux-dm.git commit 2c945820cab96ebf265598d998e63ef22393d0d4
Author: Mikulas Patocka <mpatocka@redhat.com>
Date:   Mon Mar 3 17:19:22 2014 -0500

    dm snapshot: fix metadata corruption

    Commit 55494bf2947dccdf2 ("dm snapshot: use dm-bufio") broke snapshots.
    Before that 3.14-rc1 commit, loading a snapshot's list of exceptions
    involved reading exception areas one by one into ps->area and inserting
    those exceptions into the hash table.  Commit 55494bf2947dccdf2 changed
    it so that dm-bufio with prefetch is used to load exceptions in batchs.
    Exceptions are loaded correctly, but ps->area is left uninitialized.
    When a new exception is allocated, it is stored in this uninitialized
    ps->area which will be written to the disk.  This causes metadata
    corruption.

    Fix this corruption by copying the last area that was read via dm-bufio
    into ps->area.

    Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
    Signed-off-by: Mike Snitzer <snitzer@redhat.com>

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>