[fs] fix the performance of reading /proc/mounts and friends
Message-id: <20140307041344.GD15769@sh-el6.eng.rdu2.redhat.com>
Patchwork-id: 78001
O-Subject: [rhel7][bz1072461] fix the performance of reading /proc/mounts and friends
Bugzilla: 1072461
RH-Acked-by: Ian Kent <ikent@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Every read(2) on /proc/*/mounts or /proc/*/mountinfo calls m_start() to
skip the vfsmounts we'd already read.  Unfortunately, that's done by
skipping the first N entries of a cyclic list, which means that end-to-end
read of any of those files is O(mounts^2) - we have (file size/buffer size)
read() calls, which is linear by the number of mounts and the average such
call ends up skipping half of the list entries.  And yes, on large namespaces
(thousands of mounts) it becomes visible on profiles.

It's very easy to fix - almost all such calls will have no list changes
since the previous read on the same file, and we can easily detect that,
since namespace bears an event counter.  Furthermore, they end up asking
either for the last entry we'd reached or the one right after that.
Making m_start() cache the last entry/index/event counter and avoid
re-walking the list in those cases helps a lot.

This one produces a false positive from KABI shite^Wblack magic^Wbox;
it *is* a false positive - struct mnt_namespace declaration is not
seen outside of a few files in fs/*.c.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>