[fs] Revert: Fix race when checking i_size on direct i/o read
Message-id: <531A0416.2090306@redhat.com>
Patchwork-id: 78025
O-Subject: [PATCH RHEL7] Revert: [fs] Fix race when checking i_size on direct i/o read
Bugzilla: 1071134
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

This is for bug:
Bug 1071134 - [ext4/xfstests generic/210] AIO read of last block in file returned 1 bytes, expected 300

which is a regression, causing a read of the last block of a
file via AIO to return EINVAL (not 1, as the bug summary
and test reports; *that* was a flaw in the test reporting)

The problem is, this upstream patch allows direct IO to
proceed at or beyond i_size, whereas prior to this patch,
we fell back to buffered in this case.  In RHEL7, when we
do a 4k AIO read of a 300 byte file, we first properly read
the first 300 bytes, but then come back for another DIO read
at a 300 byte offset thanks to the AIO retry loop, and that
2nd read returns EINVAL due to the unaligned IO.

Upstream, AIO doesn't do the retry, and it all seems to work
out fine, but that's after a lot of AIO reworking that's not
in RHEL7.

The patch in question DID seem to fix racing DIO readers &
writers at EOF, but AFAICT that's a long-standing bug, and
the inability to do AIO at EOF seems like a more problematic
outcome for now.

I've talked with Steven about this, and he (grudgingly ;)) agreed
to revert for now, and we'll work on a more tested solution to
both cases in a later update kernel.

Thanks,
-Eric

---

Signed-off-by: Jarod Wilson <jarod@redhat.com>