[drm] i915: fix potential oops in early irqs
Message-id: <1394216898-14417-1-git-send-email-rclark@redhat.com>
Patchwork-id: 78026
O-Subject: [PATCH] drm/i915: fix potential opps in early irqs
Bugzilla: 1073136
RH-Acked-by: Adam Jackson <ajax@redhat.com>
RH-Acked-by: Jerome Glisse <jglisse@redhat.com>

RH BZ: 1073136

irqs must be enabled early (for gmbus, dp aux transfers, etc).  But
error irqs are not expected so early and can lead to null pointer
deref if we get an error irq before the crtcs are created.  This is an
unexpected scenario, which should never happen if BIOS is doing the
right thing.  But that is not always the case.

Signed-off-by: Rob Clark <rclark@redhat.com>
---
NOTE: I've also attached the proposed upstream fix to the bz, but
at this stage I think this is the less risky solution for rhel7.
(For upstream I recommend the same patch, but with the addition of
a WARN_ON(!crtc)).

brew build: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=7158103

Signed-off-by: Jarod Wilson <jarod@redhat.com>