[powerpc] pseries: Fix parsing of initial node path in update_dt_node
Message-id: <20140310130648.7636.64292.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 78061
O-Subject: [PATCH RHEL7.0 BZ1064745 04/11] powerpc/pseries: Fix parsing of initial node path in update_dt_node
Bugzilla: 1064745
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1064745

Description:
------------
On the first call to ibm,update-properties for a node the first property
returned is the full node path. Currently this is not parsed correctly by the
update_dt_node function. Commit 2e9b7b0 attempted to fix this, but was
incorrect as it made a wrong assumption about the layout of the first
property in the work area. Further, if ibm,update-properties must be called
multiple times for the same node this special property should only be skipped
after the initial call. The first property descriptor returned consists of
the property name, property value length, and property value. The property
name is an empty string, property length is encoded in 4 byte integer, and
the property value is the node path.

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>
Acked-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
7.0

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=c8f5a57c62fcf821c435fc657f430a64205f3e99

----------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------
--
cgit v0.9.2

Signed-off-by: Jarod Wilson <jarod@redhat.com>