[x86] fix x86 fixup_irqs() error handling
Message-id: <1394475464-25009-1-git-send-email-prarit@redhat.com>
Patchwork-id: 78078
O-Subject: [RHEL7.0 PATCH BZ 1074644] x86, fix x86 fixup_irqs() error handling [v2]
Bugzilla: 1074644
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1074644

On a system with more than 256 non-system, non-per-cpu IRQs allocated, it is
possible that cpu softplug events fail silently when IRQs are "orphaned" on the
down'd cpu.  That is, there are no free vectors on the remaining cpus for the
IRQ to be assigned to.  I've pushed this patch upstream to at least warn us of
when this occurs.  This *could be* a BUG() situation, however, there are
occasions where the network fails and the system can be recovered.  Even though
it is late in RHEL7, I think this patch is safe enough in that it only is
isolated to the cpu down path and it warns of a bug in a common test that our
partners perform (disabling and enabling cpus).

Upstream: Reviewed but not in tree yet,
http://marc.info/?l=linux-kernel&m=139411150510331&w=2

Date: Mon Mar 10 14:02:08 EDT 2014
Build OS: Red Hat Enterprise Linux Workstation release 7.0 Beta (Maipo)
System name: intel-lizardhead-02.lab.bos.redhat.com with -j64
Built on: kernel-3.10.0-107.el7
Arch built: x86_64

fixup_irqs() calls chip->set_irq_affinity which eventually calls
__assign_irq_vector().  Errors are not propogated back from this function call
and this results in silent irq relocation failures.  This patch fixes this
issue and prints out a warning if there is a relocation failure.

[v2]: modified WARN to pr_crit and changed message

	Cc: Thomas Gleixner <tglx@linutronix.de>
	Cc: Ingo Molnar <mingo@redhat.com>
	Cc: "H. Peter Anvin" <hpa@zytor.com>
	Cc: x86@kernel.org
	Cc: Rui Wang <rui.y.wang@intel.com>
	Cc: Liu Ping Fan <kernelfans@gmail.com>
	Cc: Bjorn Helgaas <bhelgaas@google.com>
	Cc: Yoshihiro YUNOMAE <yoshihiro.yunomae.ez@hitachi.com>
	Cc: Lv Zheng <lv.zheng@intel.com>
	Cc: Seiji Aguchi <seiji.aguchi@hds.com>
	Cc: Yang Zhang <yang.z.zhang@Intel.com>
	Cc: Andi Kleen <ak@linux.intel.com>
	Cc: "Steven Rostedt (Red Hat)" <rostedt@goodmis.org>
	Cc: Li Fei <fei.li@intel.com>
	Cc: gong.chen@linux.intel.com
	Signed-off-by: Prarit Bhargava <prarit@redhat.com>

Cc: Igor Mammedov <imammedo@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: Tony Camuso <tcamuso@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>