[scsi] vmw_pvscsi: Fix pvscsi_abort() function
Message-id: <1394569528-25923-2-git-send-email-emilne@redhat.com>
Patchwork-id: 78187
O-Subject: [RHEL7.1 PATCH e-stor 1/3] VMW_PVSCSI: Fix pvscsi_abort() function.
Bugzilla: 1075090
RH-Acked-by: David Milburn <dmilburn@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

This change ensures that pvscsi_abort() function returns SUCCESS
only when the command in question was actually completed, otherwise
returns FAILURE. The code before change, was causing a bug where
driver tries to complete a command to the mid-layer while the mid-layer
has already requested the driver to abort that command, in response
to which the driver has responded with SUCCESS causing mid-layer
to free the command struct.

Signed-off-by: Arvind Kumar <arvindkumar@vmware.com>
Signed-off-by: Ewan Milne <emilne@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>