[redhat] Kconfig: disable infiniband ocrdma module build
Message-id: <0088e48364978f7e37db2fd3b0efd9139143318d.1395757789.git.dledford@redhat.com>
Patchwork-id: 78591
O-Subject: [Patch RHEL7] ib/ocrdma: disable module build
Bugzilla: 1079492
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jay Fenlason <fenlason@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bugzilla
--
  Bug 1079492 - blacklist or remove tech-preview ocrdma driver in RHEL 7.0 GA
  https://bugzilla.redhat.com/show_bug.cgi?id=1079492

Brew Build
--
  None.  However, we are not enabling any new code with this patch, we are
  merely disabling this driver.  I did run a make rh-configs-prep and then
  checked for the presence of the OCRDMA config option in each config and
  was able to verify that the driver is in fact turned off across all
  config files after this change.

Upstream Status
--
  Driver has been fixed upstream, but it's too late to do that here.

Testing Status
--
  None.  Emulex has verified that the be2net crash goes away when the
  ocrdma driver is never loaded at all, so we know it resolves their
  problem.  There really isn't any testing to do other than that.

A problem has been found in the ocrdma driver.  Specifically, even
though this driver is tech preview, just the act of loading this
driver makes it impossible to unload its parent be2net driver without
crashing the machine.  It's too late to fix the ocrdma driver, and
it's tech preview, so the safest thing to do at this point is to just
not build the ocrdma driver until 7.1 when we will have time to
update the driver to resolve the issue.

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>